requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1

# ML Pipeline
mlflow==2.8.1
kubeflow-pipelines==2.0.5
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
joblib==1.3.2

# Distributed Systems
raft-py==0.2.0
asyncio==3.4.3
websockets==12.0

# AI Chatbot
openai==1.3.7
transformers==4.36.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.350

# Monitoring & Telemetry
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1

# Frontend & UI
streamlit==1.28.2
plotly==5.17.0
dash==2.16.1

# Infrastructure
kubernetes==28.1.0
docker==6.1.3
redis==5.0.1
psycopg2-binary==2.9.9

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
httpx==0.25.2
